requirements.txt
python-binance==0.7.9
sqlalchemy==1.3.23
schedule==1.0.0
apprise==0.9.1
Flask==1.1.2
gunicorn==20.0.4
flask-cors==3.0.10
flask-socketio==5.0.1
eventlet==0.30.2
python-socketio[client]==5.0.4
cachetools==4.2.1
sqlitedict==1.7.0
pymongo
mongoengine
